// src/types/product.ts
export type ProductClassification = 'dietary_supplement' | 'medicine' | 'combo'

export type GetProductType = 'product_combo' | 'product_product'

export interface GetProductByAttributeMinorParams extends QueryList {
  attribute_id?: number
  attribute_value_ids?: number[]
}

export interface GetProductByCategoryParams extends QueryList {
  category_id?: number
  product_type?: GetProductType[]
}

export interface CompanyIdName {
  company_id: number
  company_name: string
}

export type ProductfilterSortType =
  | 'sold_quantity'
  | 'price_increase'
  | 'price_decrease'
  | 'new_product'
  | undefined

export interface FilterProductParams extends QueryList {
  product_type?: GetProductType
  category_ids?: number[] | []
  category_minor_ids?: number[] | []
  attributes?: any //list object update later
  attribute_id?: number
  attribute_value_ids?: number[] | []
  sort_by?: ProductfilterSortType
  price_min?: number | undefined
  price_max?: number | undefined
}

export interface ProductDetailRes {
  descendants_structor: Category[]
  product_data: ProductDetail
}

export interface Product {
  company_id: number
  product_id: number
  combo_id: number //combo
  category_id: Category
  product_code: string
  combo_code: string //combo
  product_name: string
  combo_name: string //combo
  representation_image: ImageId
  attachment_cloud_id: URLRes //combo
  image_ids: ImageId[]
  price_unit: number
  origin_price_unit: number
  sold_quantity: number
  uom_id: ProductUom
  create_date: Date
  attribute_minor_ids: AttributeMinor[]
  rel_uom_ids: ProductUom[]
  packaging_specifications: string
  promotion_category: string
  attribute_ids: ProductAttribute[]
  rel_attribute_ids: ProductRelAttribute[]
  has_variant: boolean
  star_rating: StarRatingRange
  rating_count: number
  product_available: number
  quantity: number
  price_list: PriceList
  stock_quantity: ProductUom
  product_type: ProductClassification
}

export interface PriceList {
  [productId: string]: {
    [uomId: string]: number
  }
}

export interface ProductUom {
  uom_id: number
  uom_name: string
  uom_full_standard_name?: string
  factor?: number
  quantity?: number
}

export interface AttributeId {
  attribute_id: number
  attribute_name: string
}

export interface ValueId {
  value_id: number
  value_name: string
}

export interface ProductAttribute {
  attribute: AttributeId
  attribute_value: ValueId
}

export interface ProductRelAttribute {
  attribute_id: AttributeId
  attribute_value: ValueId[]
}

export interface ListAttributeId {
  id: number
  lst_attributes_id: Array<number>
}

export interface ProductDetail extends Product {
  rel_product_ids: Product[]
  liked: boolean
  liked_count: number
  item_count: number //combo
  item_ids: ComboItem[] //combo
}

export interface ComboItem {
  item_id: number
  product_id: Product
  uom_id: ProductUom
  quantity: number
  price_unit: number
  price_total: number
}

export interface ProductDescriptionTab {
  content_id: number
  tab_id: number
  tab_name: string
  content: string
  extra_content: string
}

export interface ProductDescriptionChild extends ProductDescription {}

export interface ProductDescription {
  category_id: number
  category_name: string
  content: string
  extra_content: string
  tab: ProductDescriptionTab[]
  child: ProductDescriptionChild[]
}